Update Resque setup to be automatic
So you don't have to extend every Resque job class yourself.

The test changes a little to work around the fact that resque caches
the list of methods on each job class when `enqueue` is called, which
is before the magic installation can happen in a test scenario. In
non-test scenarios this shouldn't be a problem because the process
running the jobs will call payload_class first.