Fix travis mysql issue
We're getting this error for all mysql configs now:
```
Mysql2::Error:
  Access denied for user 'travis'@'%' to database 'metrician_test'
```

These changes switch to use the mysql root user which I believe to be
the real fix. It also updated the travis config to use `before_install`
which is what the travis docs indicate as opposed to `before_script`,
though I think both should work.

This is based on:
https://docs.travis-ci.com/user/database-setup/#MySQL