Default prefix
The previous version of this change had Metrician metrics prefixed with
"app", but you could add a custom prefix as well. This changes it so the
"app" prefix is set as a default, and you can change it using the prefix
setter on Metrician.

This is a little simpler, and allows the user to change the "app" prefix
if they want.