Remove dependence on active_support
It wasn't explicit before, and caused an issue with capistrano. Instead
of making the dependency explicit we decided to remove it since that's
where we'd like to head in the future anyway.

In order to test this change we needed a minimal gemfile to ensure other
gems didn't add these methods. This would cause tests to fails because
gems weren't loaded, so the tests have been updated to skip themselves
if their dependencies aren't loaded. This also prints messages which can
help us make sure that the tests that are being skipped should actually
be skipped.