Fix issue with code coverage reporting
The previous change exposed an issue where the first process run during
`script/test` would fail because of code coverage. This was because that
individual test process didn't test all the necessary code (and never
will). What we really care about is the total coverage at the end of
`script/test` since different code is loaded for different
configurations, but in total they should all cover all the code.

This moves the coverage check to only run at the very end in order to
fix the issue.