Install Bundler 1.17.3 in Travis before running the tests
Why is this change needed?
--------------------------
Bundler 2.0 and greater attempts to use the version of Bundler in the lock file. If it doesn't find it, you get an error like `can't find gem bundler (>= 0.a) with executable bundle`.

How does it address the issue?
------------------------------
We don't want to update all the lock files because the Rails project used for testing still requires Bundler < 2.0, so we're just making both the latest version of Bundler and the last version of Bundler compatible with the test version of Rails available. This worked in script/setup, so it'll probably work in CI.

Any links to any relevant tickets, articles or other resources?
---------------------------------------------------------------
https://travis-ci.org/Instrumental/metrician-ruby/jobs/602005821