Test Rails 5.1.3, 5.2.0, and 6.0.0 with the newest version of each gem we support for each of those versions of Rails
Why is this change needed?
--------------------------
This is a gem for automatically instrumenting Rails and popular libraries used in Rails. We should test and make sure it works for some reasonable combination of versions.

How does it address the issue?
------------------------------
Let me count the ways!

1) Adds script/generate_gemfiles

This script automatically generates a gemfile for each valid combination of Rails, the required support gems, and the gems we automatically instrument. When run, it also provides the output that needs to be added to the .travis.yml file for those gemfiles to be run as part of the test suite.

2) Adds newly generated gemfiles for Rails 5.1.3, 5.2.0, and 6.0.0

Some gems we support (e.g. dalli/memcached) are mutually exclusive, so each version of Rails has more than one gemfile.

3) Fixes an issue with Bundler versions in Travis (and other places)

If you want to install more than one version of Bundler and you don't specify a version for both, you only get one. That's bad, since some of the older gemfiles in this repo are still using Bundler 1.x.

4) Guards against infinite recursion

Newer gemfiles have newer versions of rspec which have different loading behaviors that caused the active_record instrumentation code to load more than once. It turns out, that guard code doesn't work correctly and could cause infinite recursion. Here's the code comment, for the lazy

You might wonder why this exception-based flow control is here instead of
something more sane like
instrumented_class.method_defined?(:log_without_metrician)
The answer is that method_defined?, methods, instance_methods
and other reasonable ways of answering this question do not
work correctly in this circumstance.  They return false even
when log_without_metrician is defined on the class.  The
result is that the guard statement fails and runs this code
twice.	That makes log_without_metrician an alias of
log_with_metrician and causes infinite recursion :(

It seems like some kind of deep Ruby shenanigans avoud eigen classes and/or possibly a bug. It could maybe be improved by avoiding the use of class_eval, but we did not investigate on account of how many yaks we've already shaved this week.

5) Some test environment changes
Newer versions of sidekiq don't automatically load sidekiq/testing, so that's no specifically required.

Also, newer versions of HoneyBadger change the way they disabled data sending. That's conditionalized in the tests until we stop supporting the versions older than 3.3.1.

Any links to any relevant tickets, articles or other resources?
---------------------------------------------------------------

Co-authored-by: Alex Overbeck <alex@expectedbehavior.com>